requirements.txt
flask>=2.3.0
ultralytics>=8.0.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
pymupdf>=1.23.0
imageio>=2.31.0
gunicorn>=21.0.0
roboflow>=1.1.0
